requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.30
pytest>=8.2
httpx>=0.27
SQLAlchemy>=2.0
psycopg2-binary==2.9.9

asyncpg==0.30.0
psycopg[binary]>=3.1
PyJWT>=2.8,<3
alembic>=1.12
